Error in stabilization
There’s some error in the stabilization that causes the run to timeout
when trying to print around 25 frames. I tried to update the code to be
a “no motion” stabilizer, i.e. locking the image, but it’s not doing
that either for the image sequence it does output